+++ CHANGELOG.md
@@ -6,6 +6,11 @@ Release 5.11.0 (unreleased):
  - OpenID for Verifiable Credential Issuance:
    - In `SimpleAuthorizationService` add parameter `configurationIds` to method `credentialOfferWithAuthorizationCode`
    - Support different supported credential formats having the same scope value (as this is covered by the spec)
+ - OpenID for Verifiable Presentations:
+   - In `OpenId4VpVerifier` add constructor parameter `supportedJweEncryptionAlgorithms` to advertise in metadata, deprecating `supportedJweEncryptionAlgorithm` and `supportedJweAlgorithm`
+ - OAuth 2.0:
+   - In `SimpleAuthorizationService` offer `client_attestation_pop_signing_alg_values_supported` and `client_attestation_signing_alg_values_supported` in line with [OAuth 2.0 Attestation-Based Client Authentication](https://www.ietf.org/archive/id/draft-ietf-oauth-attestation-based-client-auth-07.html#name-authorization-server-metada)
+   - Use DPoP proofs on client calls
 
 Release 5.10.1:
  - Proximity presentations: